Renaming files to have semantic names. Working through issues. Found a critical flaw in the existing code for the training pipeline.
