+++ tkyo-drift/util/logMakeEntry.js
