+++ tkyo-drift/util/logPrintCosCLI.js
