+++ tkyo-drift/util/scalarCaptureShared.js
