+++ tkyo-drift/util/scalarCompare.js
